Use double precision by default

+++ spline_rl/utils/context_builder.py
@@ -4,5 +4,5 @@ from mushroom_rl.algorithms.policy_search.black_box_optimization.context_builder
 
 class IdentityContextBuilder(ContextBuilder):
     def __call__(self, initial_state, **episode_info):
-        return torch.tensor(initial_state, dtype=torch.float32)
+        return torch.tensor(initial_state)
